V2.0 Initial
Changes to remove cmd methods from radio/slice/equalizer.
These are replaced by adding the command issuance to the radio in
setters for the different properties of slice/radio etc.

Also moved Radio/Slice instances to operate on their own run queues
rather than the default.  KVO notifications are still processed on the
default run queue.

Parser for slice and equalizer now moved into their respective class
files so that instance variables for properties can be accessed
directly thus avoiding the setter being invoked when a command to the
radio is not necessary (for example, processing a status message from
the radio).